Splitted up decryptor.c / decryptor.h
o game.c / game.h contains decryption for all SD card decryption related
stuff (SD data, NCCH/NCSD, later CIA)
o nand.c / nand.h contains everything related to dumping/decrypting from
and injecting/encrypting to the NAND
o titlekey.c / titlekey.h contains everything related to titlekeys.

Notes: a few additional #defines and functions had to be introduced to
make the split possible this far. Check nand.h.
